[delete] #30 탭 바 라우터 삭제
- 기존에 탭 바 라우터에서 해야 한다고 생각했던 역할 (메이트 요청이 오면 전체 화면 띄우기)는 탭 바 라우터가 아닌 앱 라우터가 해야하는 것이라 생각해서 제거했습니다.
- 탭 바 라우터에서 역할 맡고 있던 탭 바 아이템 생성 로직은 Builder가 담당하게 하여 좀 더 역할 분리를 명확하게 했습니다.
