+++ SniffMeet/SniffMeet/Source/Tab/TabBarModuleBuilder.swift
@@ -11,7 +11,27 @@ final class TabBarModuleBuilder {}
 
 extension TabBarModuleBuilder {
     static func build(usingSubmodules submodules: MainTabs) -> UITabBarController {
-        let tabs = TabBarRouter.makeTabs(usingSubmodules: submodules)
+        let homeTabBarItem = UITabBarItem(
+            title: "Home",
+            image: UIImage(systemName: "house.fill"),
+            tag: 0
+        )
+        let walkTabBarItem = UITabBarItem(
+            title: "walk",
+            image: UIImage(systemName: "dog.fill"),
+            tag: 1
+        )
+        let mateTabBarItem = UITabBarItem(
+            title: "mate",
+            image: UIImage(systemName: "heart.fill"),
+            tag: 2
+        )
+
+        submodules.home.tabBarItem = homeTabBarItem
+        submodules.walk.tabBarItem = walkTabBarItem
+        submodules.mate.tabBarItem = mateTabBarItem
+
+        let tabs = (submodules.home, submodules.walk, submodules.mate)
         let tabBarController = TabBarController(tabs: tabs)
         return tabBarController
     }
